[REFACTOR/#170] notification 패키지 분리
