+++ src/main/java/com/favoriteplace/app/notification/service/FCMNotificationService.java
@@ -1,9 +1,8 @@
-package com.favoriteplace.app.service.fcm;
+package com.favoriteplace.app.notification.service;
 
 import com.favoriteplace.app.domain.Member;
 import com.favoriteplace.app.repository.LikedRallyRepository;
-import com.favoriteplace.app.service.fcm.dto.PostTokenCond;
-import com.favoriteplace.app.service.fcm.enums.TotalTopicMessage;
+import com.favoriteplace.app.notification.controller.dto.PostTokenCond;
 import com.favoriteplace.global.exception.ErrorCode;
 import com.favoriteplace.global.exception.RestApiException;
 import com.google.firebase.messaging.FirebaseMessaging;
